fix: redirect at not found page
